+++ aps/aps-dag-entities/src/lib.rs
@@ -173,9 +173,6 @@ pub use aps_photograph_procedure_templates::{
 };
 pub use aps_photograph_procedures::{self, PhotographProcedure, photograph_procedures};
 pub use aps_photographs::{self, Photograph, photographs};
-pub use aps_physical_asset_model_lifecycle_profiles::{
-    self, PhysicalAssetModelLifecycleProfile, physical_asset_model_lifecycle_profiles,
-};
 pub use aps_physical_asset_models::{self, PhysicalAssetModel, physical_asset_models};
 pub use aps_physical_assets::{self, PhysicalAsset, physical_assets};
 pub use aps_pipette_tip_models::{self, PipetteTipModel, pipette_tip_models};